% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/everything.R
\name{add_cookie_javascript}
\alias{add_cookie_javascript}
\title{Add Cookies to an Existing Shiny UI}
\usage{
add_cookie_javascript(ui)
}
\arguments{
\item{ui}{A 0- or 1-argument function defining the UI of a Shiny app, or a
\code{\link[shiny:reexports]{shiny::tagList()}}.}
}
\value{
A function that takes a request and a \code{\link[shiny:reexports]{shiny::tagList()}}that includes
the dependencies needed to handle cookies.
}
\description{
Wrap a Shiny ui in this function in order to add cookie-handling
functionality.
}
\details{
Note that this function is currently exported from this package, but the
intention is to move all cookie handling to a separate package.
}
