% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookies-package.R
\docType{package}
\name{cookies-package}
\alias{cookies}
\alias{cookies-package}
\title{cookies: Use Browser Cookies with Shiny}
\description{
Cookies are name-value pairs that are saved in your browser by a website. Cookies allow websites to persist information about the user and their use of the website. This package provides tools for working with cookies in Shiny apps.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r4ds/cookies}
  \item Report bugs at \url{https://github.com/r4ds/cookies/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

Other contributors:
\itemize{
  \item Klaus Hartl (js-cookie) [copyright holder]
  \item Fagner Brack (js-cookie) [copyright holder]
  \item John Coene (JavaScript for R chapter on cookies) [copyright holder]
}

}
\keyword{internal}
