% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookies.R
\name{cookies}
\alias{cookies}
\title{Cookies}
\usage{
cookies(who, number = 10)
}
\arguments{
\item{who}{Who ate the cookies? (Requires input)}

\item{number}{How many cookies has been eaten? (Default: 10)}
}
\description{
This function will allow you to keep track of who ate the cookies.
}
\keyword{cookies}
