% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/everything.R
\name{.extract_cookie}
\alias{.extract_cookie}
\title{Extract an Individual Cookie from a Shiny Request}
\usage{
.extract_cookie(request, cookie_name)
}
\arguments{
\item{request}{A shiny request object.}

\item{cookie_name}{The name of the individual cookie to extract.}
}
\value{
The contents of that cookie.
}
\description{
Extract an Individual Cookie from a Shiny Request
}
\keyword{internal}
