% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/everything.R
\name{.extract_cookies}
\alias{.extract_cookies}
\title{Extract All Cookies from a Shiny Request}
\usage{
.extract_cookies(request)
}
\arguments{
\item{request}{A shiny request object.}
}
\value{
All cookies in the request, as a list.
}
\description{
Extract All Cookies from a Shiny Request
}
\keyword{internal}
