% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{.is_http_only}
\alias{.is_http_only}
\title{Is this cookie HttpOnly?}
\usage{
.is_http_only(cookie_name, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{cookie_name}{The cookie to check.}

\item{session}{Shiny session in which the cookies can be found (the default
should probably always be used).}
}
\value{
A logical indicating whether \code{cookie_name} is http-only.
}
\description{
HttpOnly cookies can't be manipulated via javascript.
}
\keyword{internal}
