% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{.root_session}
\alias{.root_session}
\title{Find the main session}
\usage{
.root_session(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{A session object. Probably always use the default.}
}
\value{
The first session that isn't a "session_proxy".
}
\description{
This function escapes from a module (or submodule, etc) to find the root
session.
}
\keyword{internal}
