% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{.validate_domain}
\alias{.validate_domain}
\title{Ensure domain is valid}
\usage{
.validate_domain(domain)
}
\arguments{
\item{domain}{The host to which the cookie will be sent (including
subdomains). If this is \code{NULL} (default) the cookie will only be sent to
the host of the page where this cookie was set (not including subdomains).}
}
\value{
\code{NULL} or \code{domain} as a character.
}
\description{
Ensure domain is valid
}
\keyword{internal}
