% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{.validate_expiration}
\alias{.validate_expiration}
\title{Ensure expiration is valid}
\usage{
.validate_expiration(expiration)
}
\arguments{
\item{expiration}{Days after which the cookie should expire.}
}
\value{
\code{NULL} or the expiration as a double.
}
\description{
Ensure expiration is valid
}
\keyword{internal}
