% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{.validate_http_only}
\alias{.validate_http_only}
\title{Ensure http_only is valid}
\usage{
.validate_http_only(http_only)
}
\arguments{
\item{http_only}{Logical indicating whether the cookie should only be sent as
part of an HTTP request. When this is \code{FALSE} (default), the cookie is
accessible to JavaScript via the \code{Document.cookie} property.}
}
\value{
\code{NULL} or \code{TRUE}.
}
\description{
Ensure http_only is valid
}
\keyword{internal}
