% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{.validate_logical_scalar}
\alias{.validate_logical_scalar}
\title{Ensure the argument is a single logical value}
\usage{
.validate_logical_scalar(arg_value, arg_name)
}
\arguments{
\item{arg_value}{The value of the argument.}

\item{arg_name}{The name of the argument for error messages.}
}
\value{
\code{NULL} or \code{TRUE}.
}
\description{
Ensure the argument is a single logical value
}
\keyword{internal}
