% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{.validate_path}
\alias{.validate_path}
\title{Ensure path is valid}
\usage{
.validate_path(path)
}
\arguments{
\item{path}{The path that must exist in the requested URL for the browser to
send this cookie. Includes subdirectories.}
}
\value{
\code{NULL} or \code{path} as a character.
}
\description{
Ensure path is valid
}
\keyword{internal}
