% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{.validate_scalar}
\alias{.validate_scalar}
\title{Ensure the argument is a single logical value}
\usage{
.validate_scalar(arg_value, arg_name, arg_type = "logical")
}
\arguments{
\item{arg_value}{The value of the argument.}

\item{arg_name}{The name of the argument for error messages.}

\item{arg_type}{The expected class for error messages.}
}
\value{
The value (if it is length-1).
}
\description{
Ensure the argument is a single logical value
}
\keyword{internal}
