% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{.validate_secure_only}
\alias{.validate_secure_only}
\title{Ensure secure_only is valid}
\usage{
.validate_secure_only(secure_only)
}
\arguments{
\item{secure_only}{Logical indicating whether the cookie should only be
accessible via secure (\verb{https:}) requests (except on localhost).}
}
\value{
\code{NULL} or \code{TRUE}.
}
\description{
Ensure secure_only is valid
}
\keyword{internal}
