% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{extract_cookie}
\alias{extract_cookie}
\title{Extract an Individual Cookie from a Shiny Request}
\usage{
extract_cookie(request, cookie_name)
}
\arguments{
\item{request}{A shiny request object.}

\item{cookie_name}{The name of the individual cookie to extract.}
}
\value{
The contents of that cookie.
}
\description{
The Shiny request object includes any cookies that are available to the app.
This function extracts the value of a named cookie from that request.
}
\examples{
req <- list(HTTP_COOKIE = "cookie1=expected_value; cookie2=1; cookie3=2")
extract_cookie(req, "cookie1")
extract_cookie(req, "cookie2")
extract_cookie(list(), "cookie1")
extract_cookie(NULL, "cookie1")
}
