% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{extract_cookies}
\alias{extract_cookies}
\title{Extract All Cookies from a Shiny Request}
\usage{
extract_cookies(request)
}
\arguments{
\item{request}{A shiny request object.}
}
\value{
All cookies in the request, as a list.
}
\description{
The Shiny request object includes any cookies that are available to the app.
This function extracts those cookies as a named list.
}
\examples{
req <- list(HTTP_COOKIE = "cookie1=expected_value; cookie2=1; cookie3=2")
extract_cookies(req)
extract_cookies(list())
extract_cookies(NULL)
}
