% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{remove_cookie}
\alias{remove_cookie}
\title{Remove a Cookie}
\usage{
remove_cookie(name, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{name}{The name of the cookie. Can contain any US-ASCII characters
except for: the control character, space, a tab, or separator characters
like ( ) < > @ , ; : \\ " / [ ] ? = \{ \}.}

\item{session}{Session from which to make a child scope (the default should
almost always be used).}
}
\description{
Instruct the user's browser to remove a cookie via JavaScript.
}
