% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/everything.R
\name{set_cookie}
\alias{set_cookie}
\title{Shiny Tags to Add Cookies}
\usage{
set_cookie(contents, cookie_name, expiration = 90)
}
\arguments{
\item{contents}{The contents of the cookie. Right now this should be a single
character value.}

\item{cookie_name}{A name for the cookie. Must be a valid cookie name.}

\item{expiration}{Days after which the cookie should expire.}
}
\value{
A \code{\link[shiny:reexports]{shiny::tagList()}} that provides the HTML and javascript to set the
cookie.
}
\description{
This function generates javascript which will set a cookie in the user's
browser.
}
\details{
Note that this function is currently exported from this package, but the
intention is to move all cookie handling to a separate package.
}
