% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onload.R
\name{set_cookie_on_load}
\alias{set_cookie_on_load}
\title{Shiny Tag to Add Cookies on Page Load}
\usage{
set_cookie_on_load(cookie_name, contents, expiration = 90)
}
\arguments{
\item{cookie_name}{A name for the cookie. Must be a valid cookie name.}

\item{contents}{The contents of the cookie as a single character value.}

\item{expiration}{Days after which the cookie should expire.}
}
\value{
A \code{\link[shiny:reexports]{shiny::tagList()}} that provides the HTML and javascript to set the
cookie.
}
\description{
Generates javascript which will set a cookie in the user's browser when the
Shiny app first loads.
}
\examples{
set_cookie_on_load("my_cookie", "contents of my cookie")
set_cookie_on_load("my_cookie", "contents of my cookie", expiration = 10)
}
