% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_wrapper.R
\name{add_cookie_handlers}
\alias{add_cookie_handlers}
\title{Add Cookies to an Existing Shiny UI}
\usage{
add_cookie_handlers(ui)
}
\arguments{
\item{ui}{A 0- or 1-argument function defining the UI of a Shiny app, or a
\code{\link[shiny:reexports]{shiny::tagList()}}.}
}
\value{
A function that takes a request and a \code{\link[shiny:reexports]{shiny::tagList()}} that
includes the dependencies needed to handle cookies.
}
\description{
Wrap a Shiny ui in this function in order to add cookie-handling
functionality.
}
\examples{
str(add_cookie_handlers("example"))
}
