% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookies-package.R
\docType{package}
\name{cookies-package}
\alias{cookies}
\alias{cookies-package}
\title{cookies: Use Browser Cookies with 'shiny'}
\description{
Browser cookies are name-value pairs that are saved in a user's browser by a website. Cookies allow websites to persist information about the user and their use of the website. Here we provide tools for working with cookies in 'shiny' apps, in part by wrapping the 'js-cookie' JavaScript library \url{https://github.com/js-cookie/js-cookie}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r4ds/cookies}
  \item \url{https://r4ds.github.io/cookies/}
  \item Report bugs at \url{https://github.com/r4ds/cookies/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Klaus Hartl (js-cookie JavaScript library) [copyright holder]
  \item Fagner Brack (js-cookie JavaScript library) [copyright holder]
}

}
\keyword{internal}
