% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{.http_date}
\alias{.http_date}
\title{Convert expiration days to the expected date format}
\usage{
.http_date(expiration)
}
\arguments{
\item{expiration}{Days after which the cookie should expire.}
}
\value{
A string of the date in the required format.
}
\description{
Convert expiration days to the expected date format
}
\keyword{internal}
