% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{.validate_same_site}
\alias{.validate_same_site}
\title{Ensure same_site is valid}
\usage{
.validate_same_site(same_site, secure_only)
}
\arguments{
\item{same_site}{One of "strict", "lax" (default), or "none", indicating when
the cookie should be sent. When \code{same_site = "none"}, \code{secure_only} must be
\code{TRUE}.}

\item{secure_only}{Logical indicating whether the cookie should only be
accessible via secure (\verb{https:}) requests (except on localhost).}
}
\value{
One of \code{NULL}, "Strict", "Lax", or "None".
}
\description{
Ensure same_site is valid
}
\keyword{internal}
